telegram-bot/pyproject.toml
[tool.poetry]
name = "telegram-bot"
version = "0.1.0"
description = ""
authors = ["iviv62 <ivcho2213@gmail.com>"]
readme = "README.md"
packages = [{include = "telegram_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "13.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
